refactor: 불필요하게 나뉘어 있던 예약 대기 서비스를 예약 서비스로 병합
