+++ http/Reservation.http
@@ -5,7 +5,7 @@ host: localhost:8080
 
 ### 예약 추가 API (멤버)
 POST http://localhost:8080/reservations
-cookie: token=eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJ0MUB0MS5jb20iLCJuYW1lIjoi7J6s7KaIIiwicm9sZSI6Ik1FTUJFUiIsImV4cCI6MTcxNjU0MDc2N30.zZpmVNdJ1117vqz_FDX-IqOJo5K7aL-IwZVyrV_f9ZdfoZrZUXXeX8rwpd6WdQ0z5VtP0na4-7jbJloT94ESXA;
+cookie: token=eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJ0MUB0MS5jb20iLCJuYW1lIjoi7J6s7KaIIiwicm9sZSI6Ik1FTUJFUiIsImV4cCI6MTcxNjgyMTYwNn0.jO7QpxUmLDXPtI7boEzHclPE8vDw3e6ZFwza3OY9FBc0xNX1uinFNxfq5VKo7g0mchPsvBxTq90D-Iawdyxw8A;
 host: localhost:8080
 Content-Type: application/json
